Moved input handlers to inputhandler
