+++ src/app/graph2/inputhandler/GraphEdgeInputHandler.js
