+++ src/app/graph2/inputhandler/GraphNodeInputHandler.js
