+++ src/app/graph2/inputhandler/SelectionBoxInputHandler.js
