+++ src/app/modules/fsa2/graph/inputhandler/FSAInitialMarkerInputHandler.js
