+++ src/app/modules/fsa2/graph/inputhandler/FSANodeInputHandler.js
