+++ src/app/modules/nodegraph/graph/controller/NodeGraphController.js
@@ -2,8 +2,8 @@ import GraphController from 'graph2/controller/GraphController.js';
 
 import SafeUndoNodeGraphEventHandler from 'graph2/SafeUndoNodeGraphEventHandler.js';
 
-import { GRAPH_EVENT_NODE_EDIT_WHILE_DELETE } from 'graph2/inputs/GraphNodeInputHandler.js';
-import { GRAPH_EVENT_EDGE_EDIT_WHILE_DELETE } from 'graph2/inputs/GraphEdgeInputHandler.js';
+import { GRAPH_EVENT_NODE_EDIT_WHILE_DELETE } from 'graph2/inputhandler/GraphNodeInputHandler.js';
+import { GRAPH_EVENT_EDGE_EDIT_WHILE_DELETE } from 'graph2/inputhandler/GraphEdgeInputHandler.js';
 
 import { WARNING_LAYOUT_ID } from 'session/manager/notification/NotificationManager.js';
